updated code so adding items to pantry now check for logged user

+++ pantry/database/pantry.sql.go
@@ -128,11 +128,17 @@ func (q *Queries) GetAllItems(ctx context.Context, userID string) ([]Pantry, err
 const removeItem = `-- name: RemoveItem :one
 DELETE FROM pantry
 WHERE id = ?
+    and user_id = ?
 RETURNING id, user_id, item_name, quantity, added_at, expiry_at
 `
 
-func (q *Queries) RemoveItem(ctx context.Context, id string) (Pantry, error) {
-	row := q.db.QueryRowContext(ctx, removeItem, id)
+type RemoveItemParams struct {
+	ID     string
+	UserID string
+}
+
+func (q *Queries) RemoveItem(ctx context.Context, arg RemoveItemParams) (Pantry, error) {
+	row := q.db.QueryRowContext(ctx, removeItem, arg.ID, arg.UserID)
 	var i Pantry
 	err := row.Scan(
 		&i.ID,
